+++ src/backend/model/gaming/gaming.pri
@@ -1,6 +1,5 @@
 HEADERS += \
     $$PWD/Collection.h \
-    $$PWD/CollectionList.h \
     $$PWD/Game.h \
     $$PWD/GameAssets.h \
     $$PWD/GameList.h \
@@ -8,7 +7,6 @@ HEADERS += \
 
 SOURCES += \
     $$PWD/Collection.cpp \
-    $$PWD/CollectionList.cpp \
     $$PWD/Game.cpp \
     $$PWD/GameAssets.cpp \
     $$PWD/GameList.cpp \
