+++ tests/backend/model/model.pro
@@ -2,7 +2,6 @@ TEMPLATE = subdirs
 
 SUBDIRS += \
     collection \
-    collectionlist \
     game \
     gameassets \
     gamelist \
